完善程序基础框架；添加场景读取模块；导入CUDA和Json相关库；

+++ RayTracerWithCuda/Container/List/Array.hpp
@@ -75,7 +75,7 @@ namespace EasyContainer
 	template<typename TElement>
 	INLINE Array<TElement>::Array(ArrayInitializer Initializer)
 	{
-		m_Length = Initializer.size();
+		m_Length = static_cast<int32>(Initializer.size());
 		m_Datas = new TElement[m_Length];
 		if (m_Datas == nullptr)
 		{
